Adding log rotate code in cosa_adv_security_internal.c
